enh: move codec tests closes #11536
